move make_geometry to base SAT policy
